Applied quick fixes.

+++ common/actions/data/custom/Remove.php
@@ -66,7 +66,10 @@ class Remove extends ModelAction {
 
 		$meta = new Meta();
 
-		if( $meta->load( Yii::$app->request->post(), 'Meta' ) && $meta->validate() ) {
+		// Filter reserved keys
+		$validKey = !in_array( $ckey, [ 'data', 'attributes', 'config', 'settings' ] );
+
+		if( $validKey && $meta->load( Yii::$app->request->post(), 'Meta' ) && $meta->validate() ) {
 
 			// Remove Meta
 			$this->modelService->removeDataCustomObj( $this->model, $ckey, $meta );
